Update sentiment_scoring.py
Moved out the machine learning portion to machinelearn.py. This is for sentiment scoring only. These output graphs though were the qualitative decision makers for what 'indicator in text' fields to create for regression. i.e., I used the top features to create individual fields which indicate when that feature is present, and regress it on Retweet Count.